[GenericFilter] Support custom conditions jmix-framework/jmix_1444 [GenericFilter] Support configurations jmix-framework/jmix_1445
