+++ jmix-flowui/flowui-data/src/main/java/io/jmix/flowuidata/accesscontext/FlowuiGenericFilterModifyGlobalConfigurationContext.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2023 Haulmont.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.jmix.flowuidata.accesscontext;
+
+import io.jmix.core.accesscontext.SpecificOperationAccessContext;
+
+public class FlowuiGenericFilterModifyGlobalConfigurationContext extends SpecificOperationAccessContext {
+
+    public static final String NAME = "flowui.genericfilter.modifyGlobalConfiguration";
+
+    public FlowuiGenericFilterModifyGlobalConfigurationContext() {
+        super(NAME);
+    }
+}
