+++ jmix-flowui/flowui/src/main/java/io/jmix/flowui/action/genericfilter/GenericFilterClearValuesAction.java
@@ -18,25 +18,18 @@ package io.jmix.flowui.action.genericfilter;
 
 import com.vaadin.flow.component.HasValue;
 import com.vaadin.flow.component.icon.VaadinIcon;
-import com.vaadin.flow.shared.Registration;
 import io.jmix.core.Messages;
 import io.jmix.flowui.action.ActionType;
 import io.jmix.flowui.component.filer.FilterComponent;
-import io.jmix.flowui.component.genericfilter.GenericFilter;
-import io.jmix.flowui.component.logicalfilter.LogicalFilterComponent;
 import io.jmix.flowui.kit.component.FlowuiComponentUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import javax.annotation.Nullable;
 import java.util.List;
 
 @ActionType(GenericFilterClearValuesAction.ID)
 public class GenericFilterClearValuesAction extends GenericFilterAction<GenericFilterClearValuesAction> {
 
-    public static final String ID = "filter_clearValues";
-
-    protected Registration configurationChangeRegistration;
-    protected Registration filterComponentsChangeRegistration;
+    public static final String ID = "genericFilter_clearValues";
 
     public GenericFilterClearValuesAction() {
         this(ID);
@@ -56,54 +49,6 @@ public class GenericFilterClearValuesAction extends GenericFilterAction<GenericF
     @Autowired
     protected void setMessages(Messages messages) {
         this.text = messages.getMessage("actions.genericFilter.ClearValues");
-    }
-
-    @Override
-    protected void setTargetInternal(@Nullable GenericFilter target) {
-        super.setTargetInternal(target);
-
-        if (target != null) {
-            unbindListeners();
-            bindListeners(target);
-        }
-    }
-
-    protected void bindListeners(GenericFilter target) {
-        configurationChangeRegistration = target.addConfigurationChangeListener(this::onConfigurationChanged);
-        bindFilterComponentsChangeListener(target);
-    }
-
-    protected void bindFilterComponentsChangeListener(GenericFilter target) {
-        LogicalFilterComponent<?> rootLogicalFilterComponent = target.getCurrentConfiguration()
-                .getRootLogicalFilterComponent();
-        filterComponentsChangeRegistration = rootLogicalFilterComponent
-                .addFilterComponentsChangeListener(this::onFilterComponentsChanged);
-    }
-
-    protected void unbindListeners() {
-        if (configurationChangeRegistration != null) {
-            configurationChangeRegistration.remove();
-            configurationChangeRegistration = null;
-        }
-
-        unbindFilterComponentsChange();
-    }
-
-    protected void unbindFilterComponentsChange() {
-        if (filterComponentsChangeRegistration != null) {
-            filterComponentsChangeRegistration.remove();
-            filterComponentsChangeRegistration = null;
-        }
-    }
-
-    protected void onConfigurationChanged(GenericFilter.ConfigurationChangeEvent event) {
-        unbindFilterComponentsChange();
-        bindFilterComponentsChangeListener(event.getSource());
-        refreshState();
-    }
-
-    protected void onFilterComponentsChanged(LogicalFilterComponent.FilterComponentsChangeEvent<?> event) {
-        refreshState();
     }
 
     @Override
