+++ jmix-flowui/flowui/src/main/java/io/jmix/flowui/app/filter/condition/FilterConditionDetailView.java
@@ -27,7 +27,6 @@ public abstract class FilterConditionDetailView<E extends FilterCondition>
     protected Configuration currentConfiguration;
 
     protected FilterConditionDetailView() {
-        super();
         setShowSaveNotification(false);
     }
