+++ jmix-translations/content/io/jmix/flowui/messages.properties
@@ -260,9 +260,11 @@ io.jmix.flowui.app.filter.condition/addConditionView.conditionFilterField.placeh
 io.jmix.flowui.app.filter.condition/propertyFilterConditionDetailView.title = Property condition
 io.jmix.flowui.app.filter.condition/propertyFilterConditionDetailView.parameterNameField.helperText = You can use this attribute \
   to introduce dependencies between filter components in configuration
+io.jmix.flowui.app.filter.condition/propertyFilterConditionDetailView.defaultValue = Default value
 
 # JpqlFilterConditionDetailView
 io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.title = JPQL condition
+io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.defaultValue = Default value
 
 io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.parameterClassField.label = Parameter type
 io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.parameterClassField.String = String
