+++ jmix-translations/content/io/jmix/flowui/messages_fr.properties
@@ -160,9 +160,11 @@ io.jmix.flowui.app.filter.condition/addConditionView.conditionFilterField.placeh
 # PropertyFilterConditionDetailView
 io.jmix.flowui.app.filter.condition/propertyFilterConditionDetailView.title = Éditeur de condition de propriété
 io.jmix.flowui.app.filter.condition/propertyFilterConditionDetailView.parameterNameField.helperText = Vous pouvez utiliser cet attribut pour introduire des dépendances entre les composants de filtre dans la configuration
+io.jmix.flowui.app.filter.condition/propertyFilterConditionDetailView.defaultValue = Valeur par défaut
 
 # JpqlFilterConditionDetailView
 io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.title = Condition JPQL
+io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.defaultValue = Valeur par défaut
 
 io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.parameterClassField.label = Type de paramètre
 io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.parameterClassField.String = Chaîne de caractères
