+++ jmix-translations/content/io/jmix/flowui/messages_nl.properties
@@ -191,9 +191,11 @@ io.jmix.flowui.app.filter.condition/addConditionView.conditionFilterField.placeh
 io.jmix.flowui.app.filter.condition/propertyFilterConditionDetailView.title = Eigenschap conditie bewerken
 io.jmix.flowui.app.filter.condition/propertyFilterConditionDetailView.parameterNameField.helperText = U kunt dit kenmerk gebruiken om afhankelijkheden \
   tussen filtercomponenten in de configuratie te introduceren
+io.jmix.flowui.app.filter.condition/propertyFilterConditionDetailView.defaultValue = Standaardwaarde
 
 # JpqlFilterConditionDetailView
 io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.title = JPQL-conditie
+io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.defaultValue = Standaardwaarde
 
 io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.parameterClassField.label = Parameter type
 io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.parameterClassField.String = String
