+++ jmix-translations/content/io/jmix/flowui/messages_ru.properties
@@ -258,9 +258,11 @@ io.jmix.flowui.app.filter.condition/addConditionView.conditionFilterField.placeh
 io.jmix.flowui.app.filter.condition/propertyFilterConditionDetailView.title = Условие атрибута
 io.jmix.flowui.app.filter.condition/propertyFilterConditionDetailView.parameterNameField.helperText = Этот атрибут можно \
   использовать для введения зависимостей между компонентами фильтра в конфигурации
+io.jmix.flowui.app.filter.condition/propertyFilterConditionDetailView.defaultValue = Значение по умолчанию
 
 # JpqlFilterConditionDetailView
 io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.title = JPQL условие
+io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.defaultValue = Значение по умолчанию
 
 io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.parameterClassField.label = Тип параметра
 io.jmix.flowui.app.filter.condition/jpqlFilterConditionDetailView.parameterClassField.String = Строка
